Move ProcessWaitHandle to Interop
